Display a motd message when Insights has not yet been used
Insights represents key functionality of the RHEL operating system.
Customers are missing out on Value of Subscription when not enabled.
Thus we display a message in motd when Insights is installed but
has never been used.

We use a /etc/motd.d/ drop in file to do this. Once insights-client has
been used (and a new egg has been retrieved) we drop this message.

The message can be edited or removed in the configuration file:

    /etc/insights-client/insights-client.motd

This motd message seems to represent part of the insights-client
command line RPM packaged wrapper of InsightsClient and is less
suited to the core functionality.