Point the motd message at /dev/null instead of removing
Once insights-client has been used, we remove the message in
motd recommending use of Insights. Instead of removing the symlink
we instead point it at /dev/null.

We also need to be careful for upgrades. Upgrading the
insights-client package on a system where Insights has already
been used should not result in a motd message.