Silently attempt writing to cgroup memory limits
In the case of the system not having cgroups enabled, these
ExecStartPost commands will not cause the unit to fail (because they are
prefixed with "-"). If cgroups are enabled, they will set the limits
appropriately. Either way, no errors will be logged to the journal (such
as "No such file or directory" errors) due to the stderr redirection).

Signed-off-by: Link Dupont <link@sub-pop.net>