Bypass GPG checks after first checking if file exists
The results of `gpg_validate` are used to filter down a list to one of
only valid eggs. If an egg does not exist and BYPASS_GPG is True, then
the list will erroneously contain paths to eggs that do not exist. This
ends up causing a later call to Popen.communicate to hang indefinitely.
Moving the BYPASS_GPG check to below the file exists path avoids this.

Signed-off-by: Link Dupont <link@sub-pop.net>