Enable insights-client to check for rules results following an upload
Adds a systemd "path" unit that watches /etc/insights-client/.lastupload for modification. The unit it triggers then sleeps for 120 seconds before running --check-update.

The Also= clause in insights-client.timer causes insights-client-results.path to be enabled when insights-client.timer is. This keeps the user experience "the same". Enabling just insights-client.timer will also enable insights-client-results.path. insights-client-results.path adds a PartOf= dependency to insights-client.timer, so that it is started and stopped when insights-client.timer is started.

The new ENABLE_CHECK_RESULTS variable in sysconfig enables this same functionality for cron-based systems.