build: add --enable-auto-registration flag
This flag defaults to 'no'. In order to automatically register insights
when the system is registered with RHSM, insights-client must be built
with --enable-auto-registration. This installs two additional systemd
unit files: insights-register.path and insights-register.service.
Finally, insights-register.path must be enabled post installation.
On systems that have been configured to use this auto-registration
mechanism, in order to opt out of automatic registration, mask the path
unit prior to registration with RHSM.

Signed-off-by: Link Dupont <link@sub-pop.net>