Insights register (#120)
* systemd: add insights-register service

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: add insights-register.path unit

Signed-off-by: Link Dupont <link@sub-pop.net>

* build: add --enable-auto-registration flag

This flag defaults to 'no'. In order to automatically register insights
when the system is registered with RHSM, insights-client must be built
with --enable-auto-registration. This installs two additional systemd
unit files: insights-register.path and insights-register.service.
Finally, insights-register.path must be enabled post installation.
On systems that have been configured to use this auto-registration
mechanism, in order to opt out of automatic registration, mask the path
unit prior to registration with RHSM.

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: add insights-unregister.path unit

This unit watches for the certificate to chanage, and if the file no
longer exists, runs insights-client --unregister.

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: add preset to enable path units

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: include  --force  in unregister unit

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: remove insights-register mask after unregister

Signed-off-by: Link Dupont <link@sub-pop.net>

* systemd: start register path after unregister

Signed-off-by: Link Dupont <link@sub-pop.net>