Remove egg is file condition
The insights package imported to run a Phase does not have to be an
egg file. It can simply be a folder with a cloned copy of the
insights-core repository. This is very common in development
environments and allows to live edit and use the Core files by the
Client without rebuilding the egg by making a zip file.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>