fix(logrotate): removed unused config options
This commit removes an unused option `create`
in the logrotate config. Reading further into the documentation
shows that using `create` is ignored when used together
with `copytruncate`.
The option `sharedscripts` can also be safely removed.

This does not cause any issues since it is ignored, and
can be safely removed.

Signed-off-by: Jason Jerome <jajerome@redhat.com>