fix: Do not create /root/.gnupg/ directory by accident
* Card ID: CCT-131
* Card ID: RHEL-2480
* Card ID: RHEL-2482

GPG creates a directory $HOME/.gnupg/ every time it performs some
operation. When run under root, but not manually (e.g. via
subscription-manager Cockpit plugin), it tries to create and write to
this directory, which causes SELinux denials.

This patch utilizes the `--homedir` argument GPG supports in order to
move the home directory to a temporary directory for the time of the
transaction. After the GPG action is performed, the directory is cleaned
up.