feat: basic integration testing with tmt
Introduce a minimal setup for integration tests of insights-client.
The tests are written using pytest and an helper plugin for
insights-client and other Red Hat client tools [1].

The tests are wrapped as single "integration" test for tmt [2], which
makes it possible to run them automatically e.g. in testing-farm.

There is a simple test added to show how the setup for the tests works.

[1] https://github.com/ptoscano/pytest-client-tools/
[2] https://github.com/teemtee/tmt/

Signed-off-by: Pino Toscano <ptoscano@redhat.com>