feat: run integration tests in testing-farm
Extend the packit configuration to trigger tests on testing-farm after
the completion of the build jobs.

The tests are split for the different distributions supported; this is
due to the current limitation of testing-farm in the way additional
repositories can be supplied: as the name of a Fedora Copr repository
cannot be specified, each distribution needs to get the repository URL
for it.

Manually specify the RHEL targets/composes due to the limitations in the
packit<->testing-farm mapping of RHEL composes.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>