fix: Respect PYTHONPATH environment
If PYTHONPATH is set to non-empty value, include it when running the
egg. The Egg path is always the first to ensure the correct egg will be
run, but the rest of the paths is now included.

This is useful during development. For example, `rhsm` may only be
available on non-standard path (= not as installed package), and without
this patch it is not possible to use it.