Use logger instead of custom functions
Replaces the custom logging functions client_debug and log with proper Python logging. Because the core logging attaches a handler to the root logger, the tear_down_logging function removes the remporary initialization handler to prevent duplicate printing.

To keep the log format consistent with the core logging, copied the format string, even though it introduces a piece of logic duplication. As requested, added a line number to the format. This is going to be added to the core logging too, making it consistent again.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>