fix: Allow systemd service to restart itself on failure
* Card ID: CCT-463

Historically, insights-client has not been able to recover from failures
that happened during its automatic execution. When the Core failed for
any reason (bad code, network failure, ...), the service would stay
in a 'failed' state and would not try to recover.

A few months ago, this has caused a problem that resulted in CCT-370.
See https://access.redhat.com/solutions/7056526 for more details.

The explanation of this patch follows.

---

- `Type=exec`
  With `exec`, systemd will report failure when the service's binary
  cannot be executed. This value is recommended by systemd
  documentation [1].

- `Restart=on-failure`
  Value `no` prevented us from recovering from issues, the service would
  be marked as `failed` and wouldn't try to run the next time associated
  `.timer` fired. With this setting, the systemd timer will activate
  this service even if it was failed the next time it runs.

- `RestartSec=1h`
  Time to sleep before restarting the service.

- `StartLimitBurst=6`
  This value specifies how many times the service can be executed in
  a time interval. [2]

- `StartLimitIntervalSec=12h`
  Specifies the interval the `StartLimitBurst=` applies to. Together,
  they ensure that the service is run at most six times every 12 hours
  (to make sure we don't overrun into the next day).

[1]: https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html
[2]: https://www.freedesktop.org/software/systemd/man/latest/systemd.unit.html

---

It's not trivial to test this change. In general, you have to deploy a
system with this patch installed, register with insights-client and
leave it running for several days to verify the base functionality of
running on a scheduler was not altered.

Then, you have to make sure the Core is not run or that insights-client
fails (by e.g. misconfiguring proxy). After timer runs, the service
should be marked as `failed` and also scheduled for a restarted run
in an hour. By moving the Core back on expected path, the service
should be able to recover and upload its Archive.

The same should be done the next time the service runs on timer, but
with longer outages (to check that the service is reexecuted multiple
times).

It should also be tested with the service failing all day. Even when
failing for more than 12h, the service should be activated the
next day by the timer.