fix(test): Use custom version and release for Core
* Card ID: CCT-685

This patch addresses two different problems in how we test the Core.

First, it sets the version to something very high that will never be
achieved in a reasonable time. This ensures that integration tests which
require version check to account for different behavior work correctly
even with an egg that hasn't been built and released yet.

Second, it sets the release to '0' instead of a string. This has been
breaking the version detection, since the value is usually an integer.