docs: Undocument --conf and --logging-file
* Card ID: CCT-732

The CLI includes many flags that could be reimplemented, but `conf` and
`logging-file` are harder to support and offer little value to
customers. Instead of removing them from the CLI, we'll remove them from
the man page and document the change in the RHEL release notes, as Core
is shipped to RHEL 6+ and cannot immediately remove these flags.