fix(test): Reflect changes of Core for --checkin
* Card ID: CCT-891
* Card ID: RHINENG-13984

insights-client that's not registered should not be able to check in.

Technically, it is possible to do that. When subscription-manager is
registered and the host is pushed to Inventory, its UUID is used for
matching. insights-client will collect all canonical facts, including
its own and sub-man's UUID, and send them in the checkin request.

Inventory is able to match sub-man UUID, and doesn't care where the
request came from. The end-user experience was confusing, so in Core
3.4.25+, we ensure insights-client itself is registered.

See https://github.com/RedHatInsights/insights-core/commit/00e3d5f03ceba516722a1016020f69be92cdc4bc.