fix(test): Adding debug options to the test_common_specs
As part of the work done for CCT-1107 I have added the debugging
options for this test to be able to catch error more efficiently in
case of this test failing. We will now be able to also see why the
test was actually failing and not just the error from the spec JSON.
As a minor change I have also tried to reduce the http_timeout to make
this test functional.