fix(test): Fix archive connected tests
As part of a CCT-1237 I have edited 3 different tests that are
connected to the archive creation / upload. I have removed the parts
that are connected to inventory more than the client side and therefore
we won't have to edit the tests every time core or inventory teams
change something in future. It now tests the client side only - creation
of the archive and that the archive is able to be uploaded. I have also
added the dependecies for the bz2 compressor as the test was failing due
to missing them. I have also changed checking of display_name to fqdn.
This makes the test more reliable in CI environments by comparing
the system's fully qualified domain name (FQDN) from the inventory
instead of display_name, which can differ based on system or setup.

(cherry picked from commit 171184f0ce604bc95e02f5ecbd3eb52204727fb0)