fix(test): Fixing test_tags
test_tags was failing due to not having enough time for the changes
to show in inventory. I have added the sleep function to wait 60
seconds before trying to get the inventory data back which then
passed for me locally. This is a part of an effort for CCT-1237.

(cherry picked from commit 4baa863566f694bb7b164ba939a91fa1aed20092)