fix(test): Fix the rest of RHEL8 tests
Updated the test_no_obfuscation_on_package_version to avoid attempting
to read directories from the archive, which was causing an AttributeError.
The test now checks only regular files and supports recursively verifying
files within directories like modules.d/. This ensures all relevant package files
are validated without errors. I also fixed test_check_results by changing the order
of the actions to give it a little time to reflect the ssh change later. The
test_display_name_option now skips if the error is SSL related as that is
a server issue and not test issue. This PR also fixes the registered only locally
test. This is another part of the effort for CCT-1107.