fix(test): Fixing betelgeuse docstrings
While uploading to Polarion I have figured that some of the
docstrings we are using are wrong - I have changed the sst group
and when the test have parameters I have added :parametrized:.
I have added this field also to testimony.yml so that we are
checking for this field also. Another part of effort for CCT-1090
for el8.

(cherry picked from commit 818676de25ff357fef0b8293133b6a1dc2545f20)