fix(test): Extending duration of the tests runs for el9
Some of the architectures like ppc64le or s390x are running on
emulated systems - they are running on x86_64 with e.g. s390x
emulated. This slows the execution of the tests which then usually
fail on the timeout. Until that is solved and we have other arches
available without having to emulate them we need to extend the
duration of the test runs.

(cherry picked from commit c6fbfc305a582097f7392c42f653616e7077e2c5)