fix(test): Fixed test_check_show_results
In order for insights-client to reliably check the correct results
from Advisor we needed to change the order of action in the test.
The permission of the file is changed before registering to insights
so that there is enough time for the issue to be reflected in
Advisor.

(cherry picked from commit 92538eb0a705f29eb6800aa2aab17cde300b85df)