fix(test): Fixing some architectural differences in tests
The test_tags and test_common_specs are currently not running on
other architectures other than x86_64. This will need to be fixed
some more in the future with better assert. For that we'll need
systems not emulated but the exact architecture. Until then, as it
is being worked on, we'll skip the tests for now.