fix(test): Adding parametrized option to tests
As a hopefully last touch to the tests in an effort to upload
test runs through betelgeuse to Polarion we need an parametrized
field. It was added to the master branch in PR #415 but the changes
were not backported to other branches. As it also includes changes
that are no longer needed I have created this separate PR and not
backport to fix this field.

(cherry picked from commit 4d4b187a5c08e337d30bfa126a804e24f2df9179)