fix(test): Fix Satellite failing tests
Some tests were continuously failing when run against Satellite.
The test_set_ansible_host_info was according to related CCT card
not fixed on Satellite 6.15 and therefore should be skipped in that
version. The test_compliance_policies_option was failing due to
API communication troubles so in that case it will be also skipped.
Based on some communications the test_display_name is not applicable
on Satellite and therefore skipping it entirely if it run against it.
This is the final part of CCT-1387.