Start using TMT_TEST_DATA instead of TMT_PLAN_DATA
The TMT_TEST_DATA is meant to be used by the TMT test and those files
are directly visible next to the TMT test
(in this case systemtest/tests/integration) on the TestingFarm result
page and in the TMT html report.

The TMT_PLAN_DATA on the other hand is used by the whole TMT plan
and are not associated to any specific test. Accessing those files
is quite harder in both Testing result page and in TMT html report
and those file are meant to be used for stuff related to the plan
(provisioning or preparing the system) rather than the tests.